Refactor project into modules
